move combination name builder to NameBuilder ns
