+++ src/PrestaShopBundle/Resources/config/services/adapter/product/_specific_prices.yml
@@ -7,7 +7,7 @@ services:
     arguments:
       - '@prestashop.adapter.product.specific_price.repository.specific_price_repository'
       - '@prestashop.adapter.attribute.repository.attribute_repository'
-      - '@prestashop.core.product.combination.combination_name_builder'
+      - '@prestashop.core.product.combination.name_builder.combination_name_builder'
     tags:
       - name: tactician.handler
         command: PrestaShop\PrestaShop\Core\Domain\Product\SpecificPrice\Query\GetSpecificPriceList
