+++ src/PrestaShopBundle/Resources/config/services/core/product.yml
@@ -10,5 +10,5 @@ services:
   prestashop.core.product.combination.generator.combination_generator:
     class: PrestaShop\PrestaShop\Core\Product\Combination\Generator\CombinationGenerator
 
-  prestashop.core.product.combination.combination_name_builder:
-    class: PrestaShop\PrestaShop\Core\Product\Combination\CombinationNameBuilder
+  prestashop.core.product.combination.name_builder.combination_name_builder:
+    class: PrestaShop\PrestaShop\Core\Product\Combination\NameBuilder\CombinationNameBuilder
